transform & comparison json ops split
