+++ packages/cw-jsonfilter/src/json_ops/mod.rs
@@ -1,3 +1,4 @@
+pub mod comparison;
 pub mod existence;
 pub mod json_operations;
 pub mod logical;
